# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PhiloAPI",
    "version": "2.0.0",
    "description": "A RESTful API for Philosophical Quotes. Search, filter, and discover wisdom from the ancients to the moderns."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "rapidApiKey": {
        "type": "apiKey",
        "name": "X-RapidAPI-Proxy-Secret",
        "in": "header"
      }
    },
    "schemas": {
      "Quote": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "content": {
            "type": "string",
            "example": "I know that I know nothing."
          },
          "author": { "type": "string", "example": "Socrates" },
          "field": { "type": "string", "example": "Epistemology" },
          "era": { "type": "string", "example": "Classical Greece" },
          "source": { "type": "string", "example": "Apology" }
        }
      },
      "QuoteCollection": {
        "type": "object",
        "properties": {
          "count": { "type": "integer", "example": 100 },
          "page": { "type": "integer", "example": 1 },
          "totalPages": { "type": "integer", "example": 5 },
          "results": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Quote" }
          }
        }
      }
    }
  },
  "security": [
    { "rapidApiKey": [] }
  ],
  "paths": {
    "/quotes": {
      "get": {
        "summary": "Get a list of quotes",
        "description": "Retrieve a paginated list of philosophy quotes.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 },
            "description": "Page number"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 20 },
            "description": "Number of items per page (max 100)"
          },
          {
            "name": "author",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by Author Name"
          },
          {
            "name": "field",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by Philosophical Field"
          },
          {
            "name": "lang",
            "in": "query",
            "schema": { "type": "string", "default": "en" },
            "description": "Language code (en, es, fr, pt)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/QuoteCollection" }
              }
            }
          }
        }
      }
    },
    "/quotes/random": {
      "get": {
        "summary": "Get a random quote",
        "description": "Retrieve a single random quote, optionally filtered.",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Full-text search (content or author)"
          },
          {
            "name": "author",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by Author"
          },
          {
            "name": "era",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by Era (e.g., 'Modern', 'Ancient')"
          },
          {
            "name": "field",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by Field"
          },
          {
            "name": "lang",
            "in": "query",
            "schema": { "type": "string", "default": "en" },
            "description": "Language code"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Quote" }
              }
            }
          }
        }
      }
    },
    "/quotes/{id}": {
      "get": {
        "summary": "Get a specific quote",
        "description": "Retrieve a quote by its unique ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Quote ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Quote" }
              }
            }
          },
          "404": {
            "description": "Quote not found"
          }
        }
      }
    },
    "/authors": {
      "get": {
        "summary": "Get all authors",
        "description": "Retrieve a list of all available authors.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "string" }
                }
              }
            }
          }
        }
      }
    },
    "/fields": {
      "get": {
        "summary": "Get all fields",
        "description": "Retrieve a list of all available philosophical fields.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "string" }
                }
              }
            }
          }
        }
      }
    }
  }
}